+++ test/bones_api_test.reflection.g.dart
@@ -1,6 +1,6 @@
 //
 // GENERATED CODE - DO NOT MODIFY BY HAND!
-// BUILDER: reflection_factory/2.4.10
+// BUILDER: reflection_factory/2.5.0
 // BUILD COMMAND: dart run build_runner build
 //
 
@@ -22,7 +22,7 @@ typedef __TI<T> = TypeInfo<T>;
 typedef __PR = ParameterReflection;
 
 mixin __ReflectionMixin {
-  static final Version _version = Version.parse('2.4.10');
+  static final Version _version = Version.parse('2.5.0');
 
   Version get reflectionFactoryVersion => _version;
 
@@ -61,7 +61,7 @@ class MyInfoModule$reflection extends ClassReflection<MyInfoModule>
   }
 
   @override
-  Version get languageVersion => Version.parse('3.5.0');
+  Version get languageVersion => Version.parse('3.6.0');
 
   @override
   MyInfoModule$reflection withObject([MyInfoModule? obj]) =>
@@ -126,7 +126,7 @@ class MyInfoModule$reflection extends ClassReflection<MyInfoModule>
             this,
             MyInfoModule,
             '',
-            () => (APIRoot apiRoot) => MyInfoModule(apiRoot),
+            () => MyInfoModule.new,
             const <__PR>[__PR(__TR<APIRoot>(APIRoot), 'apiRoot', false, true)],
             null,
             null,